added embedded chart page, extracted charts from server details
