+++ src/app/app.routes.ts
@@ -31,5 +31,7 @@ export const routes: Routes = [
 
   { path: 'survey/:id', loadComponent: () => import('./pages/survey-fill/survey-fill.component').then(m => m.SurveyFillComponent)},
 
+  { path: 'embed/server/:vendor/:id/:chartname', loadComponent: () => import('./pages/embedded-server-chart/embedded-server-chart.component').then(m => m.EmbeddedServerChartComponent)},
+
   { path: '**', redirectTo: '' }
 ];
