+++ src/app/pages/embedded-server-chart/embedded-server-chart.component.html
@@ -1,0 +1,9 @@
+<div class="content">
+  <app-server-charts
+    class="w-full flex flex-wrap gap-8"
+    [serverDetails]="serverDetails"
+    [benchmarksByCategory]="benchmarksByCategory"
+    [benchmarkMeta]="benchmarkMeta"
+    [showChart]="showChart"
+  ></app-server-charts>
+</div>
