+++ src/app/pages/server-details/server-details.component.scss
@@ -3,14 +3,14 @@
   min-height: 75vh;
 }
 
-.block_half {
+:host ::ng-deep .block_half {
   @apply px-6 py-6 bg-secondary border-emerald-400 border border-solid rounded-lg flex flex-col gap-4 relative overflow-hidden;
   width: calc(50% - 1.0rem);
   min-height: 460px;
   height: 460px;
 }
 
-.block_half_taller {
+:host ::ng-deep .block_half_taller {
   @apply px-6 py-6 bg-secondary border-emerald-400 border border-solid rounded-lg flex flex-col relative overflow-hidden;
   width: calc(50% - 1.0rem);
   min-height: 615px;
@@ -18,7 +18,7 @@
   margin: auto;
 }
 
-.block_full {
+:host ::ng-deep .block_full {
   @apply px-6 py-6 bg-secondary border-emerald-400 border border-solid rounded-lg flex flex-col gap-4 relative overflow-hidden;
   width: calc(100%);
   min-height: 460px;
@@ -26,7 +26,7 @@
 
 }
 
-.block_half.open {
+:host ::ng-deep .block_half.open {
   height: unset !important;
 }
